carrito

+++ TemplateTPIntegrador/Datos/DatosProductoWS.cs
@@ -11,14 +11,14 @@ namespace Datos
         int _idCategoria;
         Guid _id;
         String _nombre;
-        int _precio;
+        double _precio;
         int _stock;
         DateTime _fechaAlta;
         DateTime? _fechaBaja;
         public DatosProductoWS()
         {
         }
-        public DatosProductoWS(Guid Id, int idCategoria, string nombre, DateTime fechaAlta, DateTime fechaBaja, int precio, int stock)
+        public DatosProductoWS(Guid Id, int idCategoria, string nombre, DateTime fechaAlta, DateTime fechaBaja, double precio, int stock)
         {
             _id = Id;
             _idCategoria = idCategoria;
@@ -32,7 +32,7 @@ namespace Datos
         public int IdCategoria { get => _idCategoria; set => _idCategoria = value; }
         public Guid Id { get => _id; set => _id = value; }
         public string Nombre { get => _nombre; set => _nombre = value; }
-        public int Precio { get => _precio; set => _precio = value; }
+        public double Precio { get => _precio; set => _precio = value; }
         public int Stock { get => _stock; set => _stock = value; }
         public DateTime FechaAlta { get => _fechaAlta; set => _fechaAlta = value; }
         public DateTime? FechaBaja { get => _fechaBaja; set => _fechaBaja = value; }
